@@ -1,6 +1,5 @@
-name: myanimations
+name: flutter_ics_homescreen
 description: A new Flutter project.
-
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
@@ -20,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.18.4 <3.0.0'
+  sdk: '>=3.0.5 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -36,9 +35,9 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  flutter_cube: ^0.1.1
-  neon_widgets: ^0.1.2
 
+  syncfusion_flutter_sliders: ^19.1.56-beta
+  intl: ^0.17.0
 
 dev_dependencies:
   flutter_test:
@@ -64,14 +63,9 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
-    - assets/car.png
-    - assets/jet/
-    - assets/ural_truck_diorama_low_poly/
-    - assets/hummer-ev-low-poly/
-    - assets/blackcar.png
-    - assets/truck3d.png
-    - assets/truck2.png
-    - assets/whitecar.png
+    - asset/mylogo.png
+    - asset/cinemabg.png
+  #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
